available for sales stock fix (#13257)
* available for sales stock fix
* be lenient if the big decimal has a fractional part

+++ misc/services/camel/de-metas-camel-externalsystems/de-metas-camel-shopware6/src/main/java/de/metas/camel/externalsystems/shopware6/api/model/stock/JsonStock.java
@@ -30,8 +30,6 @@ import lombok.Builder;
 import lombok.NonNull;
 import lombok.Value;
 
-import java.math.BigDecimal;
-
 @Value
 @Builder
 @JsonDeserialize(builder = JsonStock.JsonStockBuilder.class)
@@ -39,7 +37,7 @@ public class JsonStock
 {
 	@NonNull
 	@JsonProperty("stock")
-	BigDecimal stock;
+	Integer stock;
 
 	@JsonIgnoreProperties(ignoreUnknown = true)
 	@JsonPOJOBuilder(withPrefix = "")
